modified to supprt default virtual install.

+++ ci/02-maasdeploy.sh
@@ -183,7 +183,10 @@ sed --i "s@/home/ubuntu@$HOME@g" ./deployment.yaml
 sed --i "s@qemu+ssh://ubuntu@qemu+ssh://$USER@g" ./deployment.yaml
 
 cp ./deployment.yaml ~/.juju/
-cp ./deployconfig.yaml ~/.juju/
+
+if [ -e ./deployconfig.yaml ]; then
+    cp ./deployconfig.yaml ~/.juju/
+fi
 
 sudo maas-deployer -c deployment.yaml -d --force
